Generate renamed to index
